feat: dynamic handle table and u16 handles (v0.6 Phase 2b)
Replace the fixed 256-slot handle table with a two-level structure:
base array of 256 (inline, zero allocation) + overflow pages of 256
entries each, heap-allocated on demand. Max 4096 handles per process.

Widen Handle from u8 to u16 throughout the kernel and userspace ABI.
channel_create packing changes from <<8 to <<16. All userspace handle
wrapper types (ChannelHandle, ProcessHandle, etc.) widened to u16.
wait() takes &[u16] instead of &[u8].

16 new tests for overflow insertion, cross-level lookup/close/drain,
capacity limits, and insert_at in overflow range. 2,411 tests pass.

+++ system/libraries/protocol/init.rs
@@ -166,14 +166,15 @@ pub struct DocConfig {
     /// Byte length of the encoded image in the File Store.
     pub img_file_store_length: u32,
     /// Kernel channel handle for the editor channel.
-    pub editor_handle: u8,
+    pub editor_handle: u16,
     /// Kernel channel handle for the decoder channel.
-    pub decoder_handle: u8,
+    pub decoder_handle: u16,
     /// Kernel channel handle for the document service (filesystem) channel.
-    pub fs_handle: u8,
+    pub fs_handle: u16,
     /// Kernel channel handle for the core (presenter) channel.
-    pub core_handle: u8,
+    pub core_handle: u16,
 }
+
 const _: () = assert!(core::mem::size_of::<DocConfig>() <= 60);
 
 // ── Legacy 9P filesystem (init ↔ virtio-9p driver) ──────────────────
